Move `ReminderList` component to separate file
